Add warning message in web script for null values.
Since HiPack does not handle properly null values, in Try It
section of the web we would like to handle if the user add
a null value. In this case, the null value is ignored, but
a warning message is shown, explained how HiPack deals
with this kind of values.

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>